refactor: refactor packages that divide all the classes into 'domain' and 'view'
