+++ src/main/java/autoracing/domain/Car.java
@@ -1,4 +1,4 @@
-package autoracing;
+package autoracing.domain;
 
 import java.util.ArrayList;
 import java.util.List;
