+++ src/main/java/autoracing/domain/RandomIntGenerator.java
@@ -1,4 +1,4 @@
-package autoracing;
+package autoracing.domain;
 
 import java.util.Random;
