+++ src/test/java/autoracing/RacingGameIntegrationTest.java
@@ -1,5 +1,8 @@
 package autoracing;
 
+import autoracing.domain.RacingGame;
+import autoracing.domain.RandomRacingRule;
+import autoracing.view.InputView;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
