+++ src/test/java/autoracing/domain/RacingGameTest.java
@@ -1,4 +1,4 @@
-package autoracing;
+package autoracing.domain;
 
 import org.junit.jupiter.api.Test;
